Refactor: 기존 코드 타입 변경

See merge request s12-final/S12P31A204!5

+++ apps/electron-client/src/renderer/pages/settings/components/SettingsMailList.tsx
@@ -1,24 +1,69 @@
-import { AuthUser } from "@/types/authType";
+import { AccountsResponse } from "@/types/authType";
 
 import useAuthenticateStore from "@stores/authenticateStore";
+
+import { useDeleteAccount } from "@hooks/useGetUser";
 
 import IconButton from "@components/common/button/IconButton";
 
 import GoogleIcon from "@assets/icons/GoogleIcon";
+import NaverIcon from "@assets/icons/NaverIcon";
+import MiniSettingIcon from "@assets/icons/MiniSettingsIcon";
 import DeleteIcon from "@assets/icons/DeleteIcon";
 
-const InnerList = ({ user }: { user: AuthUser }) => {
+const InnerList = ({
+  user,
+  onEdit,
+}: {
+  user: AccountsResponse;
+  onEdit: (account: AccountsResponse | null) => void;
+}) => {
   const { deleteAuthUser } = useAuthenticateStore();
 
-  function handleDelete() {
-    deleteAuthUser(user);
+  const { mutateAsync: deleteAccount } = useDeleteAccount();
+
+  async function handleDelete() {
+    if (!user) return;
+
+    if (!confirm(`${user.email}\n계정을 삭제하시겠습니까?`)) return;
+
+    try {
+      const data = await deleteAccount({ accountId: user.id });
+
+      if (data.success) {
+        alert(`${user.email}\n계정이 삭제되었습니다!`);
+        deleteAuthUser(user); // 임시..
+      }
+    } catch (error) {
+      console.error("Error deleting account:", error);
+      alert("계정 삭제에 실패했습니다.");
+      return;
+    }
   }
+
+  function findDomain(email: string) {
+    const domain = email.split("@")[1].split(".")[0];
+
+    if (domain === "gmail") {
+      return "Gmail";
+    } else if (domain === "naver") {
+      return "Naver";
+    }
+
+    return domain.charAt(0).toUpperCase() + domain.slice(1);
+  }
+
+  const domain = findDomain(user.email);
 
   return (
     <div className="flex items-center justify-between p-2 w-full">
       <div className="flex justify-center items-center px-1 gap-3 w-fit h-fit">
         <span className="bg-blue-700 rounded-full p-2">
-          <GoogleIcon />
+          {domain === "Gmail" ? (
+            <GoogleIcon />
+          ) : (
+            <NaverIcon strokeColor="white" />
+          )}
         </span>
 
         <div className="flex flex-col items-start justify-between w-fit h-full">
@@ -27,23 +72,38 @@ const InnerList = ({ user }: { user: AuthUser }) => {
         </div>
       </div>
 
-      <IconButton
-        icon={<DeleteIcon />}
-        className="p-2 hover:bg-red-200 rounded-full transition-all duration-300"
-        onClick={() => {
-          handleDelete();
-        }}
-      />
+      <div className="flex items-center justify-center gap-0.5 w-fit h-fit">
+        <IconButton
+          icon={<DeleteIcon />}
+          className="p-2 hover:bg-red-200 rounded-full transition-all duration-300"
+          onClick={() => {
+            handleDelete();
+          }}
+        />
+        <IconButton
+          icon={<MiniSettingIcon />}
+          className="p-2 hover:bg-blue-200 rounded-full transition-all duration-300"
+          onClick={() => {
+            onEdit(user);
+          }}
+        />
+      </div>
     </div>
   );
 };
 
-const SettingsMailList = ({ users }: { users: AuthUser[] }) => {
+const SettingsMailList = ({
+  users,
+  onEdit,
+}: {
+  users: AccountsResponse[];
+  onEdit: (account: AccountsResponse | null) => void;
+}) => {
   return (
     <div className="flex flex-col justify-center items-center p-1 gap-1 rounded-2xl bg-white">
       {users.map((user, index) => (
         <span key={user.id} className="w-full h-fit">
-          <InnerList user={user} />
+          <InnerList user={user} onEdit={onEdit} />
           {users.length - 1 !== index && (
             <hr className="border-t border-light1 w-[95%]" />
           )}
